+++ src/components/layout/AppFooter/src/index.vue
@@ -3,15 +3,21 @@ defineOptions({ name: 'AppFooter' });
 </script>
 
 <template>
-  <div class="app-footer"></div>
+  <div class="app-footer">
+    © 2025 Madello AI  | Madello Consulting
+  </div>
+  
 </template>
 
 <style lang="css" scoped>
 .app-footer {
+  display: flex;
   position: fixed;
   bottom: 0;
-  width: 350vw;
+  width: 100%;
+  justify-content: center;
   z-index: -1;
+  margin: 8px;
 
   img {
     transform: translateX(-100vw);