+++ src/pages/signature/components/SignaturePanel.vue
@@ -86,7 +86,7 @@ function changeZoom(value: number) {
 }
 
 .signature-panel .btn-small {
-  text-shadow: 0 0 0 #b2248c;
+  text-shadow: 0 0 0 #4C00FF;
   color: transparent;
 }